requirements.txt
Flask==3.0.3
google-ai-generativelanguage==0.6.6
google-generativeai==0.7.2
python-dotenv==1.0.1